// backend/src/listings/types/sort-type.type.ts
export enum SortType {
  PRICE_ASC = 'price-asc',
  PRICE_DESC = 'price-desc',
  SURFACE_AREA_ASC = 'surface-area-asc',
  SURFACE_AREA_DESC = 'surface-area-desc',
  CONSTRUCTION_YEAR_ASC = 'construction-year-asc',
  CONSTRUCTION_YEAR_DESC = 'construction-year-desc',
  NEWEST = 'newest',
  OLDEST = 'oldest',
}